LPS-146207 Split customer portal layout file to specific files
