+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/layouts/BaseLayout/index.js
@@ -9,8 +9,8 @@
  * distribution rights of the Software.
  */
 
-import ProjectContacts from '../ProjectContacts';
-import QuickLinksPanel from '../QuickLinksPanel';
+import ProjectContacts from '../../components/ProjectContacts';
+import QuickLinksPanel from '../../containers/QuickLinksPanel';
 
 const Layout = ({children, hasProjectContact, hasQuickLinks, project}) => {
 	return (
